third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-weight-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'font-weight' to a number: 0
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
Harness: the test ran to completion.
